requirements.txt
pandas==1.5.3
pandas-ta
python-telegram-bot==20.7
requests
